style: change DTO models to User model
