+++ src/controller/UserController.ts
@@ -4,7 +4,8 @@ import { UserDatabase } from "../data/UserDatabase";
 import { IdGenerator } from "../services/IdGenerator";
 import { HashManager } from "../services/HashManager";
 import { Authenticator } from "../services/Authenticator";
-import { SignupInputDTO, SignupBusiness } from "../business/SingnupBusiness";
+import { SignupBusiness } from "../business/SingnupBusiness";
+import { SignupInputDTO } from "../model/User";
 
 
 export class UserController {
