+++ src/model/User.ts
@@ -60,3 +60,15 @@ export class User {
         ) 
 	}
 }
+
+export interface SignupInputDTO {
+    name: string,
+    email: string,
+    nickname: string,
+    password: string
+}
+
+export interface LoginInputDTO {
+	emailOrNickname: string;
+	password: string;
+}
